Add Bodu TOML 1.0 corpus as KAT-record tests; share tagged-JSON compare
Vendor the supplied hand-authored TOML 1.0.0 corpus (10 valid with
expected value trees, 22 invalid) consolidated into toml-test tagged-JSON
form, and drive it from a dedicated, well-defined TomlV10Kat known-answer
record in TomlV10CorpusTests: every valid vector parses under the strict
v1.0 default and matches its expected tree; every invalid vector —
including the TOML 1.1 syntax it deliberately carries (\e, \xHH, optional
seconds, multi-line / trailing-comma inline tables) — is rejected.

Extract the tagged-JSON equivalence comparison out of TomlConformanceTests
into a shared TaggedJson helper so the toml-test conformance and the v1.0
corpus reuse one implementation.
