+++ Bodu.Text.Formats/test/Bodu.Text.Formats.Test.csproj
@@ -30,6 +30,8 @@
     <EmbeddedResource Include="Text.Toml\Conformance\toml-test-invalid.json" />
     <EmbeddedResource Include="Text.Toml\Conformance\toml-test-files-1.0.0.txt" />
     <EmbeddedResource Include="Text.Toml\Conformance\toml-test-files-1.1.0.txt" />
+    <EmbeddedResource Include="Text.Toml\Conformance\bodu-toml10-valid.json" />
+    <EmbeddedResource Include="Text.Toml\Conformance\bodu-toml10-invalid.json" />
   </ItemGroup>
 
   <ItemGroup>
